+++ ios/Frameworks/KMPWorkManager.xcframework/ios-arm64/KMPWorkManager.framework/Headers/KMPWorkManager.h
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:1e04d553e826ff5db1d1bc663472ef544885b4b54f6f3d6900db80d23e1e965b
-size 297639
+oid sha256:3da4384a4558a41cbd8a2146f57e3a830a68874bbe0ba06ad774786ceb5d33f1
+size 298879